requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
playwright==1.53.0
pyee==13.0.0
PyYAML==6.0.2
requests==2.32.4
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
